+++ core/src/main/kotlin/ca/allanwang/kau/utils/FileUtils.kt
@@ -6,13 +6,14 @@ import java.io.File
 import java.io.IOException
 import java.io.InputStream
 import java.text.SimpleDateFormat
+import java.util.*
 
 /**
  * Created by Allan Wang on 2017-08-04.
  */
 @Throws(IOException::class)
 fun createMediaFile(prefix: String, extension: String): File {
-    val timeStamp = SimpleDateFormat.getDateInstance().parse("yyyyMMdd_HHmmss")
+    val timeStamp = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
     val imageFileName = "${prefix}_${timeStamp}_"
     val storageDir = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES)
     val frostDir = File(storageDir, prefix)
@@ -20,9 +21,9 @@ fun createMediaFile(prefix: String, extension: String): File {
     return File.createTempFile(imageFileName, extension, frostDir)
 }
 
-@Throws
+@Throws(IOException::class)
 fun Context.createPrivateMediaFile(prefix: String, extension: String): File {
-    val timeStamp = SimpleDateFormat.getDateInstance().parse("yyyyMMdd_HHmmss")
+    val timeStamp = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
     val imageFileName = "${prefix}_${timeStamp}_"
     val storageDir = getExternalFilesDir(Environment.DIRECTORY_PICTURES)
     return File.createTempFile(imageFileName, extension, storageDir)